Change time db integer to bigint

+++ db/migrate/20210614143231_add_restart_time_to_worktimes.rb
@@ -1,5 +1,5 @@
 class AddRestartTimeToWorktimes < ActiveRecord::Migration[6.0]
   def change
-    add_column :worktimes, :restart_time, :integer
+    add_column :worktimes, :restart_time, :bigint
   end
 end
